+++ pam/gkr-pam-stubs.c
@@ -43,7 +43,7 @@ egg_memory_unlock (void)
 }
 
 void*
-egg_memory_fallback (void *p, unsigned long sz)
+egg_memory_fallback (void *p, size_t sz)
 {
 	/* Handles allocation, reallocation and freeing */
 	return realloc (p, sz);
