+++ pkcs11/user-store/gck-user-standalone.c
@@ -39,7 +39,7 @@ void egg_memory_lock (void)
 	{ g_static_mutex_lock (&memory_mutex); }
 void egg_memory_unlock (void) 
 	{ g_static_mutex_unlock (&memory_mutex); }
-void* egg_memory_fallback (void *p, unsigned long sz) 
+void* egg_memory_fallback (void *p, size_t sz) 
 	{ return g_realloc (p, sz); }
 
 CK_RV
